plotting checkpoint
