+++ example/2. Design/2.1.0 - Special Object Propreties.py
