+++ src/limitstates/objects/section/__init__.py
@@ -1,3 +1,2 @@
 from .section import *
 from .clt import *
-from .plot import PlotDisplayProps, plotSection
